RavenDB-18442 - Move another static methods to BackupUtils
