+++ src/Raven.Server/Utils/BackupUtils.cs
@@ -1,16 +1,132 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using NCrontab.Advanced;
 using Raven.Client.Documents.Operations.Backups;
 using Raven.Client.Documents.Operations.OngoingTasks;
+using Raven.Client.Json.Serialization;
+using Raven.Client.ServerWide.Operations;
 using Raven.Client.Util;
 using Raven.Server.Documents;
+using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.NotificationCenter.Notifications;
+using Raven.Server.ServerWide;
+using Raven.Server.ServerWide.Context;
+using Sparrow.LowMemory;
 
 namespace Raven.Server.Utils;
 
 internal static class BackupUtils
 {
+    internal static BackupInfo GetBackupInfo(GetBackupInfoParameters parameters)
+    {
+        var oneTimeBackupStatus = GetBackupStatusFromCluster(parameters.ServerStore, parameters.Context, parameters.DatabaseName, taskId: 0L);
+
+        if (parameters.PeriodicBackups.Count == 0 && oneTimeBackupStatus == null)
+            return null;
+
+        var lastBackup = 0L;
+        PeriodicBackupStatus lastBackupStatus = null;
+        var intervalUntilNextBackupInSec = long.MaxValue;
+        if (oneTimeBackupStatus?.LastFullBackup != null && oneTimeBackupStatus.LastFullBackup.Value.Ticks > lastBackup)
+        {
+            lastBackup = oneTimeBackupStatus.LastFullBackup.Value.Ticks;
+            lastBackupStatus = oneTimeBackupStatus;
+        }
+
+        foreach (var periodicBackup in parameters.PeriodicBackups)
+        {
+            var status = ComparePeriodicBackupStatus(periodicBackup.Configuration.TaskId,
+                backupStatus: GetBackupStatusFromCluster(parameters.ServerStore, parameters.Context, parameters.DatabaseName, periodicBackup.Configuration.TaskId),
+                inMemoryBackupStatus: periodicBackup.BackupStatus);
+
+            if (status.LastFullBackup != null && status.LastFullBackup.Value.Ticks > lastBackup)
+            {
+                lastBackup = status.LastFullBackup.Value.Ticks;
+                lastBackupStatus = status;
+            }
+
+            if (status.LastIncrementalBackup != null && status.LastIncrementalBackup.Value.Ticks > lastBackup)
+            {
+                lastBackup = status.LastIncrementalBackup.Value.Ticks;
+                lastBackupStatus = status;
+            }
+
+            var nextBackup = BackupUtils.GetNextBackupDetails(new BackupUtils.NextBackupDetailsParameters
+            {
+                OnParsingError = null, // For skipping error messages notification and log
+                OnMissingNextBackupInfo = null, // For skipping error messages notification and log
+                Configuration = periodicBackup.Configuration,
+                BackupStatus = status,
+                ResponsibleNodeTag = parameters.ServerStore.NodeTag,
+                NodeTag = parameters.ServerStore.NodeTag
+            });
+            if (nextBackup == null)
+                continue;
+
+            if (nextBackup.TimeSpan.Ticks < intervalUntilNextBackupInSec)
+                intervalUntilNextBackupInSec = nextBackup.TimeSpan.Ticks;
+        }
+
+        return new BackupInfo
+        {
+            LastBackup = lastBackup == 0L ? (DateTime?)null : new DateTime(lastBackup),
+            IntervalUntilNextBackupInSec = intervalUntilNextBackupInSec == long.MaxValue ? 0 : new TimeSpan(intervalUntilNextBackupInSec).TotalSeconds,
+            BackupTaskType = lastBackupStatus?.TaskId == 0 ? BackupTaskType.OneTime : BackupTaskType.Periodic,
+            Destinations = AddDestinations(lastBackupStatus)
+        };
+    }
+
+    internal static PeriodicBackupStatus GetBackupStatusFromCluster(ServerStore serverStore, TransactionOperationContext context, string databaseName, long taskId)
+    {
+        var statusBlittable = serverStore.Cluster.Read(context, PeriodicBackupStatus.GenerateItemName(databaseName, taskId));
+
+        if (statusBlittable == null)
+            return null;
+
+        var periodicBackupStatusJson = JsonDeserializationClient.PeriodicBackupStatus(statusBlittable);
+        return periodicBackupStatusJson;
+    }
+
+    internal static PeriodicBackupStatus ComparePeriodicBackupStatus(long taskId, PeriodicBackupStatus backupStatus, PeriodicBackupStatus inMemoryBackupStatus)
+    {
+        if (backupStatus == null)
+        {
+            backupStatus = inMemoryBackupStatus ?? new PeriodicBackupStatus { TaskId = taskId };
+        }
+        else if (inMemoryBackupStatus?.Version > backupStatus.Version && inMemoryBackupStatus.NodeTag == backupStatus.NodeTag)
+        {
+            // the in memory backup status is more updated
+            // and is of the same node (current one)
+            backupStatus = inMemoryBackupStatus;
+        }
+
+        return backupStatus;
+    }
+
+    private static List<string> AddDestinations(PeriodicBackupStatus backupStatus)
+    {
+        if (backupStatus == null)
+            return null;
+
+        var destinations = new List<string>();
+        if (backupStatus.UploadToAzure?.Skipped == false)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.Azure));
+        if (backupStatus.UploadToGlacier?.Skipped == false)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.AmazonGlacier));
+        if (backupStatus.UploadToFtp?.Skipped == false)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.FTP));
+        if (backupStatus.UploadToGoogleCloud?.Skipped == false)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.GoogleCloud));
+        if (backupStatus.UploadToS3?.Skipped == false)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.AmazonS3));
+        if (backupStatus.LocalBackup?.TempFolderUsed == false)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.Local));
+        if (destinations.Count == 0)
+            destinations.Add(nameof(BackupConfiguration.BackupDestination.None));
+
+        return destinations;
+    }
 
     public static NextBackup GetNextBackupDetails(NextBackupDetailsParameters parameters)
     {
@@ -22,7 +138,7 @@ internal static class BackupUtils
             BackupFrequency = parameters.Configuration.FullBackupFrequency,
             Configuration = parameters.Configuration,
             LastBackupUtc = lastFullBackupUtc,
-            OnParsingError = parameters.OnParsingError
+            OnParsingError = parameters.OnParsingError,
         });
         var nextIncrementalBackup = GetNextBackupOccurrence(new GetNextBackupOccurrenceParameters
         {
@@ -34,24 +150,7 @@ internal static class BackupUtils
 
         if (nextFullBackup == null && nextIncrementalBackup == null)
         {
-            parameters.OnMissingNextBackupInfo?.Invoke();
-            
-            // if (database != null)
-            // {
-            //     var message = "Couldn't schedule next backup " +
-            //                   $"full backup frequency: {configuration.FullBackupFrequency}, " +
-            //                   $"incremental backup frequency: {configuration.IncrementalBackupFrequency}";
-            //     if (string.IsNullOrWhiteSpace(configuration.Name) == false)
-            //         message += $", backup name: {configuration.Name}";
-            //
-            //     database.NotificationCenter.Add(AlertRaised.Create(
-            //         database.Name,
-            //         "Couldn't schedule next backup, this shouldn't happen",
-            //         message,
-            //         AlertType.PeriodicBackup,
-            //         NotificationSeverity.Warning));
-            // }
-
+            parameters.OnMissingNextBackupInfo?.Invoke(parameters.Configuration);
             return null;
         }
 
@@ -110,7 +209,6 @@ internal static class BackupUtils
                 Exception = e,
                 BackupFrequency = parameters.BackupFrequency,
                 Configuration = parameters.Configuration
-                //TODO-v: e, parameters.BackupFrequency, parameters.Configuration 
             });
 
             return null;
@@ -155,6 +253,33 @@ internal static class BackupUtils
                (nextIncrementalBackup == null || nextFullBackup <= nextIncrementalBackup);
     }
 
+    internal static void CheckServerHealthBeforeBackup(ServerStore serverStore, string name)
+    {
+        if (serverStore.Server.CpuCreditsBalance.BackgroundTasksAlertRaised.IsRaised())
+        {
+            throw new BackupDelayException($"Failed to start Backup Task: '{name}'. The task cannot run because the CPU credits allocated to this machine are nearing exhaustion.")
+            {
+                DelayPeriod = serverStore.Configuration.Server.CpuCreditsExhaustionBackupDelay.AsTimeSpan
+            };
+        }
+
+        if (LowMemoryNotification.Instance.LowMemoryState)
+        {
+            throw new BackupDelayException($"Failed to start Backup Task: '{name}'. The task cannot run because the server is in low memory state.")
+            {
+                DelayPeriod = serverStore.Configuration.Backup.LowMemoryBackupDelay.AsTimeSpan
+            };
+        }
+
+        if (LowMemoryNotification.Instance.DirtyMemoryState.IsHighDirty)
+        {
+            throw new BackupDelayException($"Failed to start Backup Task: '{name}'. The task cannot run because the server is in high dirty memory state.")
+            {
+                DelayPeriod = serverStore.Configuration.Backup.LowMemoryBackupDelay.AsTimeSpan
+            };
+        }
+    }
+
     public class GetNextBackupOccurrenceParameters
     {
         public string BackupFrequency { get; set; }
@@ -189,8 +314,14 @@ internal static class BackupUtils
 
         public Action<OnParsingErrorParameters> OnParsingError { get; set; }
 
-        public OnMissingNextBackupInfo OnMissingNextBackupInfo { get; set; }
+        public Action<PeriodicBackupConfiguration> OnMissingNextBackupInfo { get; set; }
     }
 
-    internal delegate void OnMissingNextBackupInfo(); // Action with no parameters
+    public class GetBackupInfoParameters
+    {
+        public TransactionOperationContext Context { get; set; }
+        public ServerStore ServerStore { get; set; }
+        public List<PeriodicBackup> PeriodicBackups { get; set; }
+        public string DatabaseName { get; set; }
+    }
 }
