+++ src/Raven.Server/Web/System/DatabasesHandler.cs
@@ -422,7 +422,7 @@ namespace Raven.Server.Web.System
                             periodicBackups.Add(new PeriodicBackup
                             {
                                 Configuration = periodicBackupConfiguration,
-                                BackupStatus = PeriodicBackupRunner.GetBackupStatusFromCluster(ServerStore, context, databaseName, periodicBackupConfiguration.TaskId)
+                                BackupStatus = BackupUtils.GetBackupStatusFromCluster(ServerStore, context, databaseName, periodicBackupConfiguration.TaskId)
                             });
                         }
 
@@ -434,7 +434,15 @@ namespace Raven.Server.Web.System
                             [nameof(DatabaseInfo.NodesTopology)] = nodesTopology.ToJson(),
                             [nameof(DatabaseInfo.DeletionInProgress)] = DynamicJsonValue.Convert(dbRecord.DeletionInProgress),
                             [nameof(DatabaseInfo.Environment)] = studioEnvironment,
-                            [nameof(DatabaseInfo.BackupInfo)] = PeriodicBackupRunner.GetBackupInfoLogic(context, ServerStore, periodicBackups, databaseName)
+                            [nameof(DatabaseInfo.BackupInfo)] = BackupUtils.GetBackupInfo(
+                                new BackupUtils.GetBackupInfoParameters()
+                                {
+                                    ServerStore = ServerStore,
+                                    PeriodicBackups = periodicBackups,
+                                    DatabaseName = databaseName,
+                                    Context = context
+                                    //context, ServerStore, periodicBackups, databaseName
+                                })
                         };
 
                         context.Write(writer, databaseInfoJson);
